add build tag to security tests
